+++ src/routes/TweetList.jsx
@@ -1,16 +1,48 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import ErrorMessages from "../components/atoms/message/ErrorMessages";
 import TweetForm from "../components/molecules/TweetForm";
 import TweetItems from "../components/molecules/TweetItems";
+import { getTweets } from "../apis/tweet";
 
 const TweetList = () => {
   const [errorMessages, setErrorMessages] = useState([]);
   const [addTweet, setAddTweet] = useState(null);
+  const [items, setItems] = useState([]);
+  const [hasMore, setHasMore] = useState(true);
+  const [offset, setOffset] = useState(0);
+
+  useEffect(() => {
+    fetchMoreData();
+  }, []);
+
+  const fetchMoreData = async () => {
+    const response = await getTweets(offset);
+    if (response.success) {
+      const newItems = response.data;
+
+      if (newItems.length === 0) {
+        setHasMore(false);
+        return;
+      }
+      setItems([...items, ...newItems]);
+      setOffset(offset + newItems.length);
+    } else {
+      setHasMore(false);
+      setErrorMessages(response.errors);
+    }
+  };
 
   const handleAddTweet = (tweet, previews) => {
     const newTweet = { ...tweet, images: previews };
     setAddTweet(newTweet);
   };
+
+  useEffect(() => {
+    if (addTweet) {
+      setItems([addTweet, ...items]);
+      setOffset(offset + 1);
+    }
+  }, [addTweet]);
 
   return (
     <div className="flex-1 max-w-screen-sm">
@@ -22,7 +54,11 @@ const TweetList = () => {
         setErrorMessages={setErrorMessages}
         onAddTweet={handleAddTweet}
       />
-      <TweetItems setErrorMessages={setErrorMessages} addTweet={addTweet} />
+      <TweetItems
+        items={items}
+        hasMore={hasMore}
+        fetchMoreData={fetchMoreData}
+      />
     </div>
   );
 };
